requirements.txt
pymupdf
pytesseract
opencv-python
pillow